more tidying
